Serialize and deserialize a struct with a single integer
Implemented the logic for serialization:
Until Reflection is used, the struct to be serialized shall contain a
tuple of member pointers to the member fields to be serialized.
This tuple is used for both serialization and deserialization.
The order of the members in the tuple set the field numbers in the proto.